requirements.txt
requests~=2.26.0
PyDrive~=1.3.1
httplib2~=0.18.1
oauth2client~=4.1.3
google-api-python-client~=2.30.0
pandas~=1.3.4
future~=0.18.2
pytz~=2021.3
mysql-connector-python~=8.0.27
DateTime~=4.3
selenium~=4.1.0
termcolor~=1.1.0
telegram~=0.0.1